Api/develop

See merge request s12-s-project/S12P21S004!191

+++ api/src/main/java/com/in4sight/api/dto/TimeSeriesDataDto.java
@@ -63,7 +63,7 @@ public class TimeSeriesDataDto {
 	@NoArgsConstructor
 	@AllArgsConstructor
 	private static class Threshold {
-		private int warning;
-		private int critical;
+		private Double warning;
+		private Double critical;
 	}
 }
